+++ packages/devtools_shared/lib/src/server/handlers/_devtools_extensions.dart
@@ -11,7 +11,7 @@ abstract class _ExtensionsApiHandler {
     ServerApi api,
     Map<String, String> queryParams,
     ExtensionsManager extensionsManager,
-    DTDConnectionInfo? dtd,
+    DTDInfo? dtd,
   ) async {
     final missingRequiredParams = ServerApi._checkRequiredParameters(
       [ExtensionsApi.packageRootUriPropertyName],